+++ tests/StreamTest.php
@@ -70,10 +70,6 @@ class StreamTest extends TestCase
 
     public function testConvertsToStringNonSeekableStream(): void
     {
-        if (defined('HHVM_VERSION')) {
-            self::markTestSkipped('This does not work on HHVM.');
-        }
-
         $handle = popen('echo foo', 'r');
         $stream = new Stream($handle);
         self::assertFalse($stream->isSeekable());
@@ -82,10 +78,6 @@ class StreamTest extends TestCase
 
     public function testConvertsToStringNonSeekablePartiallyReadStream(): void
     {
-        if (defined('HHVM_VERSION')) {
-            self::markTestSkipped('This does not work on HHVM.');
-        }
-
         $handle = popen('echo bar', 'r');
         $stream = new Stream($handle);
         $firstLetter = $stream->read(1);